+++ src/main/java/org/folio/inventory/resources/InstancesBatch.java
@@ -20,6 +20,7 @@ import io.vertx.ext.web.handler.BodyHandler;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.lang3.tuple.Pair;
 import org.folio.inventory.common.WebContext;
+import org.folio.inventory.consortium.services.ConsortiumService;
 import org.folio.inventory.domain.BatchResult;
 import org.folio.inventory.domain.instances.Instance;
 import org.folio.inventory.domain.instances.titles.PrecedingSucceedingTitle;
@@ -39,8 +40,8 @@ public class InstancesBatch extends AbstractInstances {
   public static final String BATCH_RESPONSE_FIELD_ERROR_MESSAGES = "errorMessages";
   public static final String BATCH_RESPONSE_FIELD_TOTAL_RECORDS = "totalRecords";
 
-  public InstancesBatch(final Storage storage, final HttpClient client) {
-    super(storage, client);
+  public InstancesBatch(final Storage storage, final HttpClient client, final ConsortiumService consortiumService) {
+    super(storage, client, consortiumService);
   }
 
   public void register(Router router) {